lms/requirements.txt
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.11.2
black==23.1.0
bootstrap4==0.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
crispy-bootstrap4==2022.1
cryptography==39.0.1
cssutils==2.6.0
decorator==5.1.1
defusedxml==0.7.1
Django==3.2.16
django-baton==2.6.2
django-colorfield==0.8.0
django-crispy-forms==2.0
django-extensions==3.2.1
django-inlinecss-redux==0.4.0
django-location-field==2.1.0
django-phonenumber-field==7.0.2
django-suit==0.2.26
django_debug_toolbar==3.8.1
executing==1.2.0
Faker==17.0.0
flake8==6.0.0
idna==3.4
ipython==8.10.0
jedi==0.18.2
marshmallow==3.19.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.0
parso==0.8.3
pathspec==0.11.0
pexpect==4.8.0
phonenumberslite==8.13.6
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
prompt-toolkit==3.0.36
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
Pygments==2.14.0
PyJWT==2.6.0
pynliner==0.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
soupsieve==2.4
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.9.0
urllib3==1.26.14
wcwidth==0.2.6
webargs==8.2.0
